Refactor map components
- Remove eslintrc-lenient
- Remove eslint disable
- Replace react app on storybook dir with appropriate stories
- Separate out styles
- Move out common components
