--- storybook/.eslintrc-lenient.json
@@ -1,28 +1,0 @@
-{
-  "extends": [
-    "react-app",
-    "react-app/jest"
-  ],
-  "rules": {
-    "import/no-relative-packages": 0,
-    "import/extensions": ["off", "never"],
-    "jsx-a11y/click-events-have-key-events": "warn",
-    "jsx-a11y/no-static-element-interactions": "warn",
-    "no-param-reassign": "warn",
-    "consistent-return": "warn",
-    "no-underscore-dangle": "warn",
-    "react/jsx-filename-extension": ["error", {
-      "extensions": [".js", ".jsx", ".ts", ".tsx"]
-    }]
-  },
-  "overrides": [
-    {
-      "files": [
-        "**/*.stories.*"
-      ],
-      "rules": {
-        "import/no-anonymous-default-export": "off"
-      }
-    }
-  ]
-}
