+++ storybook/.storybook/preview.ts
@@ -1,4 +1,5 @@
 import type { Preview } from "@storybook/react";
+
 import '@the-deep/reporting-module-components/build/esm/index.css';
 import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
