+++ storybook/README.md
@@ -31,6 +31,4 @@ npm run storybook
 ## Run linter
 ```bash
 npm run lint
-# Or
-npm run lint:lenient
 ```
