+++ storybook/package.json
@@ -1,6 +1,5 @@
 {
-  "name": "reporting",
-  "homepage": "http://matthewsmawfield.github.io/map-vizzard",
+  "name": "@the-deep/reporting-module-storybook",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -41,17 +40,12 @@
     "webfontloader": "^1.6.28"
   },
   "scripts": {
-    "start": "DISABLE_ESLINT_PLUGIN=true react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
     "lint": "eslint src --ext .js,.jsx,.ts,.tsx",
-    "lint:lenient": "eslint src --ext .js,.jsx,.ts,.tsx --config .eslintrc-lenient.json",
     "storybook": "storybook dev -p 6006 --no-open",
     "build-storybook": "storybook build",
     "deploy-storybook": "storybook-to-ghpages",
     "predeploy": "npm run build",
-    "deploy": "DISABLE_ESLINT_PLUGIN=true gh-pages -d build"
+    "deploy": "gh-pages -d build"
   },
   "browserslist": {
     "production": [
