--- storybook/src/App.module.css
@@ -1,8 +1,0 @@
-.App {
-  /* min-width: 600px; */
-  min-height: 50px;
-  width: 100%;
-  height: 100vh;
-  overflow: hidden;
-  overflow-y: scroll;
-}
