+++ storybook/src/components/ColorScale/styles.module.css
