+++ storybook/src/components/Map/Layers/HexbinLayer.jsx
@@ -10,8 +10,7 @@ import {
 import { fromLonLat } from 'ol/proj';
 import * as d3ColorScale from 'd3-scale-chromatic';
 import { scaleLog } from 'd3-scale';
-// import { rgba } from '../../MapVizzard/MapOptions/ColorPicker';
-import { vector } from '../Source';
+import { vector } from '../helpers';
 
 function HexbinLayer({
   map,
