+++ storybook/src/components/Map/Layers/SymbolLayer.jsx
@@ -13,10 +13,7 @@ import { unByKey } from 'ol/Observable';
 import Feature from 'ol/Feature';
 import Point from 'ol/geom/Point';
 import { fromLonLat } from 'ol/proj';
-// import { rgba } from '../../MapVizzard/MapOptions/ColorPicker';
-import { numberFormatter } from '../Helpers';
-import { vector } from '../Source';
-import { rgba } from '../../MapVizzard/MapOptions/ColorPicker';
+import { numberFormatter, vector, rgba } from '../helpers';
 import capital from '../assets/map-icons/capital.svg';
 import city from '../assets/map-icons/city.svg';
 import settlement from '../assets/map-icons/settlement.svg';
