--- storybook/src/components/Map/Layers/index.jsx
@@ -1,12 +1,0 @@
-import VectorLayer from './VectorLayer';
-import LineLayer from './LineLayer';
-import TileLayer from './TileLayer';
-import MapboxLayer from './MapboxLayer';
-import MaskLayer from './MaskLayer';
-import SymbolLayer from './SymbolLayer';
-import HeatmapLayer from './HeatmapLayer';
-import HexbinLayer from './HexbinLayer';
-
-export {
-  VectorLayer, LineLayer, MapboxLayer, TileLayer, MaskLayer, SymbolLayer, HeatmapLayer, HexbinLayer,
-};
