+++ storybook/src/components/Map/OlMap/styles.module.css
@@ -1,0 +1,84 @@
+.mapBottomRight {
+  top: unset !important;
+  left: unset !important;
+  right: 8px;
+  bottom: 8px !important;
+  color: inherit !important;
+  position: absolute;
+  z-index: 100;
+  text-align: right;
+}
+
+.mapBottomLeft {
+  top: unset !important;
+  left: 8px !important;
+  right: unset !important;
+  bottom: 20px !important;
+  color: inherit !important;
+  position: absolute;
+  z-index: 100;
+  text-align: left;
+}
+
+.mapTopRight {
+  top: 64px !important;
+  left: unset !important;
+  right: 8px;
+  bottom: unset !important;
+  color: inherit !important;
+  position: absolute;
+  z-index: 100;
+  text-align: right;
+}
+
+.mapTopLeft {
+  top: 64px !important;
+  left: 8px !important;
+  right: unset !important;
+  bottom: unset !important;
+  color: inherit !important;
+  position: absolute;
+  z-index: 100;
+  text-align: left;
+}
+
+/* FIXME: headeriMMAP is not available here */
+.headeriMMAP .POS-topRight {
+  top: 36px !important;
+}
+
+.POS-bottomRight {
+  top: unset !important;
+  left: unset !important;
+  right: 8px;
+  bottom: 8px !important;
+  color: inherit !important;
+  opacity: 0.95;
+}
+
+.POS-topRight {
+  bottom: unset !important;
+  left: unset !important;
+  top: 8px;
+  right: 9px;
+  opacity: 0.95;
+  color: inherit !important;
+}
+
+.POS-bottomLeft {
+  left: 8px !important;
+  top: unset !important;
+  right: unset !important;
+  bottom: 8px;
+  opacity: 0.8;
+  color: inherit !important;
+}
+
+@media only screen and (max-width: 600px) {
+  /* FIXME: headeriMMAP is not available here */
+  .headeriMMAP .POS-topRight {
+    top: 10px !important;
+    zoom: 0.7;
+  }
+}
+
