--- storybook/src/components/Map/Source/index.js
@@ -1,8 +1,0 @@
-import vector from './vector';
-import xyz from './xyz';
-import osm from './osm';
-import mask from './mask';
-
-export {
-  vector, xyz, osm, mask,
-};
