--- storybook/src/components/Map/Source/mask.js
@@ -1,7 +1,0 @@
-import { Vector as VectorSource } from 'ol/source';
-
-function mask() {
-  return new VectorSource({ wrapX: false });
-}
-
-export default mask;
