--- storybook/src/components/Map/Source/osm.js
@@ -1,7 +1,0 @@
-import * as olSource from 'ol/source';
-
-function osm() {
-  return new olSource.OSM();
-}
-
-export default osm;
