+++ storybook/src/components/Map/styles.module.css
@@ -1,7 +1,7 @@
-/* @import url('https://fonts.googleapis.com/css2?family=Alegreya:wght@400;700&family=Barlow:wght@400;700&family=Futura:wght@400;700&family=Lato:wght@400;700&family=Montserrat:wght@400;700&family=Open Sans:wght@400;700&family=Raleway:wght@400;700&family=Roboto:wght@400;700&family=RobotoSlab:wght@400;700&display=swap'); */
-/* @import url('https://fonts.googleapis.com/css2?family=Alegreya:wght@400;700&family=Lato:wght@400;700&family=Montserrat:wght@400;700&family=Roboto:ital,wght@0,100;0,300;0,400;0,700;1,300;1,400&display=swap'); */
+/* FIXME: Remove this */
 @import url('https://fonts.googleapis.com/css2?family=Alegreya:wght@400;700&family=Barlow+Condensed:wght@400;700&family=Barlow:wght@100;400;700&family=Lato:wght@400;700&family=Montserrat:wght@400;700&family=Open+Sans:wght@400;700&family=Raleway:wght@400;700&family=Roboto+Slab:wght@400;700&family=Roboto:ital,wght@0,100;0,300;0,400;0,700;1,300;1,400&display=swap');
 
+/* FIXME: remove ol-* */
 .ol-map {
   min-width: 700px;
   min-height: 50px;
@@ -46,17 +46,6 @@
 
 .ol-overviewmap-box {
   pointer-events: none !important;
-}
-
-.mapContainer {
-  -webkit-box-shadow: 0px 0px 11px -8px rgba(0,0,0,0.45);
-  -moz-box-shadow: 0px 0px 11px -8px rgba(0,0,0,0.45);
-  box-shadow: 0px 0px 11px -8px rgba(0,0,0,0.45);
-  -webkit-user-select: none; /* Safari */
-  -ms-user-select: none; /* IE 10 and IE 11 */
-  user-select: none; /* Standard syntax */
-  background-color: white;
-  position: relative;
 }
 
 .logos {
@@ -265,91 +254,6 @@
   height: 12px;
 }
 
-.headeriMMAP .scalePostopRight, 
-.headeriMMAP .POS-topRight {
-  top: 36px !important;
-}
-
-.mapBottomRight {
-  top: unset !important;
-  left: unset !important;
-  right: 8px;
-  bottom: 8px !important;
-  color: inherit !important;
-  position: absolute;
-  z-index: 100;
-  text-align: right;
-}
-
-.mapBottomLeft {
-  top: unset !important;
-  left: 8px !important;
-  right: unset !important;
-  bottom: 20px !important;
-  color: inherit !important;
-  position: absolute;
-  z-index: 100;
-  text-align: left;
-}
-
-.mapTopRight {
-  top: 64px !important;
-  left: unset !important;
-  right: 8px;
-  bottom: unset !important;
-  color: inherit !important;
-  position: absolute;
-  z-index: 100;
-  text-align: right;
-}
-
-.mapTopLeft {
-  top: 64px !important;
-  left: 8px !important;
-  right: unset !important;
-  bottom: unset !important;
-  color: inherit !important;
-  position: absolute;
-  z-index: 100;
-  text-align: left;
-}
-
-.scalePosbottomRight {
-  top: unset !important;
-  left: unset !important;
-  right: 8px;
-  bottom: 8px !important;
-  color: inherit !important;
-  opacity: 0.8;
-}
-
-.POS-bottomRight {
-  top: unset !important;
-  left: unset !important;
-  right: 8px;
-  bottom: 8px !important;
-  color: inherit !important;
-  opacity: 0.95;
-}
-
-.scalePostopRight, .POS-topRight {
-  bottom: unset !important;
-  left: unset !important;
-  top: 8px;
-  right: 9px;
-  opacity: 0.95;
-  color: inherit !important;
-}
-
-.scalePosbottomLeft, .POS-bottomLeft {
-  left: 8px !important;
-  top: unset !important;
-  right: unset !important;
-  bottom: 8px;
-  opacity: 0.8;
-  color: inherit !important;
-}
-
 .mapFooter {
   font-size: 9px;
   padding-left: 8px;
@@ -439,10 +343,6 @@
   display: inline;
 }
 
-.embedMap {
-  box-shadow: none;
-}
-
 .exportButtons {
   margin-top: 8px;
   transform: scale(0.8);
@@ -469,7 +369,7 @@
     padding-top: 0px !important;
     margin-bottom: 50px !important;
   }
-  .headeriMMAP .POS-topRight {
+  .headeriMMAP {
     top: 10px !important;
     zoom: 0.7;
   }
