--- storybook/src/index.css
@@ -1,35 +1,0 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-.MuiTabs-indicator {
-  height: 5px !important;
-  background-color: #00000030 !important;
-}
-
-header a {
-  font-family: 'Barlow Condensed' !important;
-}
-
-.tooltip {
-  /* border-radius: 0px; */
-  border: 0.5px solid rgb(192, 192, 192);
-  background-color: rgb(255, 255, 255);
-  padding: 2px 4px;
-  font-size: 12px;
-  line-height: 1.13;
-}
-
-.tooltip b {
-  font-size: 16px;
-}
