--- storybook/src/stories/MapOptionsMapbox.stories.js
@@ -1,19 +1,0 @@
-import MapOptions from '../components/MapVizzard/MapOptions';
-import mapConfig from './sudan.json';
-
-// Story Config
-export default {
-  title: 'Map Vizzard/Layer options panel/Mapbox',
-  component: MapOptions,
-  tags: ['autodocs'],
-  argTypes: {
-    children: { table: { disable: true } },
-  },
-};
-
-export const Default = {
-  args: {
-    layers: mapConfig.layers,
-    activeLayer: 11,
-  },
-};
