+++ storybook/src/stories/default.json
@@ -1,0 +1,44 @@
+{
+  "mapOptions": {
+    "height": 512,
+    "width": 654,
+    "showHeader": true,
+    "mainTitle": "Default map",
+    "subTitle": "Country level",
+    "dateText": "",
+    "center": { "lon": 30.1, "lat": 48.4 },
+    "zoom": 3,
+    "maxZoom": 7,
+    "minZoom": 1,
+    "showScale": true,
+    "showLogos": [],
+    "scaleUnits": "metric",
+    "scaleBar": false,
+    "scaleBarPosition": "bottomRight",
+    "enableMouseWheelZoom": false,
+    "enableDoubleClickZoom": false,
+    "enableZoomControls": true,
+    "zoomControlsPosition": "topRight",
+    "showLegend": true,
+    "legendPosition": "bottomLeft",
+    "showFooter": true,
+    "sources": "My Source",
+    "fontStyle": {
+      "fontWeight": "normal",
+      "fontFamily": "Roboto Slab",
+      "color": {
+        "r": 0, "g": 0, "b": 0, "a": 1
+      }
+    }
+  },
+  "layers": [
+    {
+      "id": 1,
+      "name": "OSM background",
+      "type": "osm",
+      "opacity": 1,
+      "zIndex": 1,
+      "visible": 1
+    }
+  ]
+}
