+++ storybook/tsconfig.json
@@ -20,7 +20,5 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": [
-    "src"
-, "../lib/src/Timeline"  ]
+  "include": ["src"]
 }
